refactor: rewrite cjsToEsm.cjs to it's equivalent in ECMAScript notation
